Store Spotify token in session
Store SpotifyToken and expiration date time in session so that login is
not required everytime a user load the page again.
Expiration time is checked if the token is still valid, otherwise a new
login is required.

Added 'moment' as dependency

Fixed async call in componentWillUmount() which cause warnings in
console